Short-circuit libs and header lookups.
Sue me, I cheat it, but wanted this to work, will figure out all the
logic in a future refactoring.

In the meantime, you can use --disable-lookup to force the extension
take both --with-iconv-dir and --with-freetds-dir that you can supply.

This helps the cross compilation as LIBRARY_PATH (part of
MiniPortile#activate) is useless in cross compilation environments.