Fixed timeout with new non-blocking mode.
With FreeTDS, timeouts happen via sending a message/error to the error
handling indicating the timeout has been hit. With the change to not
immediately throwing an exception due to the non-blocking mode of the
DB requests, the code that cancelled an in-progress query was not
called.

This commit adds a call to dbcancel() when an error occurs and adds a
regression test to ensure this is not broken in the future.