Refactor build of freetds and iconv recipes and add OpenSSL.
extconf.rb now checks for availability of freetds, with a fallback to the
bundled mini_portile recipes. As a further extension, it might be useful
to include the stable freetds sources into the gem.

This allows native build on Windows with download and compilation of all
dependent libraries. This also adds a appveyor config for ci-testing of
the native build. However not all tests pass, so far.

For building windows binary gems this makes the switch to the new rake-compiler-dock.

The mini_portile recipes are changed, now, to build shared libraries. This
decreases the fat binary Windows gem from around 7 MB to 2.5 MB, compared to a
static build. To avoid any version mismatchs with OpenSSL, I used a more specific
naming scheme for OpenSSL dlls. This should avoid any contact with the DLL hell.

This fixes issues #200, #75 and hopefully helps for #196.