Add compat with RubyInstaller-2.4.1
It adds gem metadata for automatic installation of the freetds package
which is part of the official MSYS2/MINGW repository now. This makes
installation of the tiny_tds source gem as easy as installation of the
binary windows gem, but works for RubyInstaller-2.4+ only.
The freetds MINGW package is maintained here:
https://github.com/Alexpux/MINGW-packages/tree/master/mingw-w64-freetds

The Appveyor tests still work with our ports build and doesn't test
against the new freetds package, so far.

This disables the perl hack for OpenSSL, when the perl version is
recent enough. Otherwise mixing git's shell commands into the path breaks
MSYS2 builds, as used by RubyInstaller-2.4.

This also removes lib directories from DLL search path, because they
don't contain any DLLs.

The update of libiconv is required for compat with MSYS2.

The patch file is derived from
https://github.com/Alexpux/MINGW-packages/tree/master/mingw-w64-freetds