starting on ruby 3.1 support
clearly, something changes in ruby 3.1 and tls. freetds receives an
error "Adaptive Server connection failed", which turns out to be
ssl related. updating dependencies version did not fix the issue. but
further research found it is a known issue with other projects that
depend on the freetds library.

i tried updating openssl to the latest version, that didn't work.
the only thing that has given progress so far is setting freetds
to build with gnutls instead of openssl, which seem to be solution
accepted by most people.

as a result, it may be needed to switch tinytds to gnutls