CI improvements (#525)
* Lock Bundler to v2.3.26

This is the last version that supports Ruby 2.4 and 2.5.

* Remove fix to use rake-compiler-dock on Windows for libiconv

The original author states that the code I removed is for using virtualbox on Windows to compile the gem. I read up on this and while rake-compiler-dock still supports using docker-machine to compile gems, the project has been discontinued in favor of Docker Desktop, which nowadays only works with Docker in WSL2.

* Bump openssl to v1.1.1s

1.1.1d gives an error when compiling against i686-w64-mingw32. Upgrading it seems to work.

* Define task to compile native extensions

* Skip dead connection test on Windows

* Extend pipeline to test gem on Windows

* Package gems on CI before Windows tests

* Ensure test container uses Ruby version defined in CI matrix

* Test against Ruby 2.4

* Publish test results to CircleCI

* Rename artifact for gems

* Restore cross-compiled code from final gem file

* Store precompiled gem into workspace instead of cache

this seems to be the more appropriate type of data store according to the CircleCI documentation.

* Fix order for compiling ports

* Disable progress bar for Invoke-WebRequest

* Add `/ALLUSERS` option to RubyInstaller

Looks like a new option that has been introduced in the most recent 2.7 version and explains the timeout we see with the Ruby installation on CircleCI.
https://github.com/oneclick/rubyinstaller2/wiki/FAQ#user-content-install-mode

Older versions seem to ignore the parameter and run the installation as usual.