Fix dangling pointer after client.close (#527)
* Test(client): Add dead test after closing client

We recently encountered an unexpected Segmentation Fault Error when
checking whether a client was dead after closing it.

To consider this in the future, I've updated the test cases for the
client to include this scenario, so we can catch any issues earlier on.

* Fix(client) Set dangling pointer to NULL after client closure

We recently encountered an unexpected Segmentation Fault Error while
checking if a client was dead after closing it. Upon investigation, we
discovered that the issue was caused by deallocating all memory used by
the client structure without setting the pointer to NULL after closing
the client connection through dbclose. This resulted in a dangling
pointer, which led to the Segmentation Fault when we tried to check
whether the client was dead using TDS dbdead.

To prevent this issue from occurring again, I have made the necessary
updates to set the client pointer to NULL after every dbclose. This will
ensure that we avoid any dangling pointers and keep our code running
smoothly.

---------

Co-authored-by: Matias Martini <martini@fintual.com>