Mark `alloc` as undefined for `TinyTds::Result`
From my understanding, you only need to define an `alloc` function if you want to define a C struct and store values on it. For example, `TinyTds::Client` defines such an `allocate` function to wrap its client. On `TinyTds::Result`, it does not seem we do have a struct and therefore does not need an `alloc` function.