Don't run native test for Ruby 2.4 on Windows
I tried couple of things in order to get msys2 to run with Ruby 2.4, going so far that I even manually installed msys2. However, this also failed, but I found a StackOverflow thread about it:

https://stackoverflow.com/questions/73985340/error-target-not-found-when-trying-to-install-msys2-and-mingw-development-toolc

But this seems to much to me: If we need to go such lengths in order to test a "minor" thing for one Ruby version, I rather don't do it.